+++ network_test.go
@@ -27,7 +27,8 @@ func TestFactorRow(t *testing.T) {
 		},
 	}
 
-	net := New("test", []Variable{a, b, c}, []Factor{factor})
+	net, err := New("test", "", []Variable{a, b, c}, []Factor{factor})
+	assert.Nil(t, err)
 
 	variable := net.Variables()[2]
 
@@ -71,7 +72,8 @@ func TestNetworkToVE(t *testing.T) {
 		}},
 	}
 
-	n := New("umbrella", vars, factors)
+	n, err := New("umbrella", "", vars, factors)
+	assert.Nil(t, err)
 
 	v, variables, err := n.toVE(nil)
 	assert.Nil(t, err)
@@ -122,7 +124,8 @@ func TestNetworkSolveUmbrella(t *testing.T) {
 		}},
 	}
 
-	n := New("umbrella", vars, factors)
+	n, err := New("umbrella", "", vars, factors)
+	assert.Nil(t, err)
 	policy, err := n.SolvePolicies(true)
 	assert.Nil(t, err)
 
@@ -208,7 +211,8 @@ func TestNetworkSolveOil(t *testing.T) {
 		}},
 	}
 
-	n := New("oil", vars, factors)
+	n, err := New("oil", "", vars, factors)
+	assert.Nil(t, err)
 	policy, err := n.SolvePolicies(true)
 	assert.Nil(t, err)
 
@@ -272,7 +276,8 @@ func TestNetworkRearrange(t *testing.T) {
 		Given: []string{"a", "b", "c"},
 	}
 
-	net := New("test", variables, []Factor{fac})
+	net, err := New("test", "", variables, []Factor{fac})
+	assert.Nil(t, err)
 
 	_, f, err := net.SolveQuery(map[string]string{}, []string{"a", "b", "c", "d"}, true)
 	assert.Nil(t, err)